+++ jenkins-client/src/main/java/com/offbytwo/jenkins/model/BuildResult.java
@@ -7,10 +7,26 @@
 package com.offbytwo.jenkins.model;
 
 public enum BuildResult {
-    FAILURE, UNSTABLE, REBUILDING, BUILDING, ABORTED, SUCCESS, UNKNOWN, NOT_BUILT, 
+    FAILURE, UNSTABLE, REBUILDING, BUILDING,
     /**
-     * This will be the result of a job in cases where it has been 
-     * cancelled during the time in the queue. 
+     * This means a job was already running and has been aborted.
+     */
+    ABORTED,
+    /**
+     * 
+     */
+    SUCCESS,
+    /**
+     * ?
+     */
+    UNKNOWN,
+    /**
+     * This is returned if a job has never been built.
+     */
+    NOT_BUILT,
+    /**
+     * This will be the result of a job in cases where it has been cancelled
+     * during the time in the queue.
      */
     CANCELLED
 }