+++ jenkins-client/src/main/java/com/offbytwo/jenkins/model/TestReport.java
@@ -1,5 +1,6 @@
 package com.offbytwo.jenkins.model;
 
+import java.util.Collections;
 import java.util.List;
 
 /**
@@ -8,10 +9,37 @@ import java.util.List;
  */
 public class TestReport extends BaseModel {
 
+    public static final String EMPTY_STRING = "";
+
+    /**
+     * This will be returned by the API in cases where the build has not been
+     * run.
+     */
+    public static final TestReport NO_TEST_REPORT = new TestReport(0, 0, 0, EMPTY_STRING,
+            Collections.<TestChildReport>emptyList());
+
     private int failCount;
     private int skipCount;
     private int totalCount;
     private String urlName;
+
+    private TestReport(int failCount, int skipCount, int totalCount, String urlName,
+            List<TestChildReport> childReports) {
+        super();
+        this.failCount = failCount;
+        this.skipCount = skipCount;
+        this.totalCount = totalCount;
+        this.urlName = urlName;
+        this.childReports = childReports;
+    }
+
+    public TestReport() {
+        this.failCount = 0;
+        this.skipCount = 0;
+        this.totalCount = 0;
+        // FIXME: What is the best choice to initialize?
+        this.urlName = EMPTY_STRING;
+    }
 
     private List<TestChildReport> childReports;